Move the base from index.js to lib/base.js
Future implementations, e.g. via ssh, can build upon this class
